requirements.txt
faiss-cpu==1.6.5
FlagEmbedding==1.2.10
huggingface-hub==0.23.3
numpy==1.24.2
pyarrow==16.1.0
pyarrow-hotfix==0.6
safetensors==0.4.3
scipy==1.13.1
sentence-transformers==3.0.1
tokenizers==0.19.1
torch==2.0.0
torchvision==0.15.1
tqdm==4.66.4
transformers==4.41.2
